make offersearch inner class
